+++ packages/experiments/src/components/Stack/Stack.types.ts
@@ -1,23 +1,15 @@
 import { IStyle } from '../../Styling';
-import { IStatelessComponent, IStyleableComponentProps } from '../../Foundation';
+import { IStatelessComponent, IStyleableComponentProps, IStylesProp } from '../../Foundation';
 
 export type Alignment = 'start' | 'end' | 'center' | 'space-between' | 'space-around' | 'space-evenly' | 'baseline' | 'stretch';
 
-export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
-
 export type IStackComponent = IStatelessComponent<IStackProps, IStackStyles>;
-
-// contains the members of IStackProps that are common to both VerticalStack and HorizontalStack
-export type IPartialStackProps = Omit<
-  IStackProps,
-  'verticalAlignment' | 'horizontalAlignment' | 'horizontal' | 'verticalGap' | 'horizontalGap' | 'wrap' | 'styles'
->;
 
 export interface IStackProps extends IStyleableComponentProps<IStackProps, IStackStyles>, React.HTMLAttributes<HTMLElement> {
   /**
    * How to render the Stack.
    */
-  as?: string | React.ReactType<IPartialStackProps>;
+  as?: string | React.ReactType<IStackProps>;
 
   /**
    * CSS class name used to style the Stack.
@@ -31,6 +23,7 @@ export interface IStackProps extends IStyleableComponentProps<IStackProps, IStac
 
   /**
    * Whether to render Stack child elements horizontally.
+   * @defaultvalue false
    */
   horizontal?: boolean;
 
@@ -42,7 +35,7 @@ export interface IStackProps extends IStyleableComponentProps<IStackProps, IStac
   /**
    * How to align Stack child elements vertically (along the y-axis).
    */
-  verticalAlignment?: Alignment;
+  verticalAlignment?: 'top' | 'bottom' | Alignment;
 
   /**
    * Whether the Stack should take up 100% of the width of its parent.
@@ -70,6 +63,11 @@ export interface IStackProps extends IStyleableComponentProps<IStackProps, IStac
   gap?: number | string;
 
   /**
+   * Vertical gap between Stack child elements.
+   */
+  verticalGap?: number | string;
+
+  /**
    * Maximum width of the Stack.
    */
   maxWidth?: number | string;
@@ -86,18 +84,14 @@ export interface IStackProps extends IStyleableComponentProps<IStackProps, IStac
 
   /**
    * Whether Stack children should wrap onto multiple rows (HorizontalStacks only).
+   * @defaultvalue false
    */
   wrap?: boolean;
 
   /**
-   * Horizontal gap between Stack child elements.
+   * Custom styles to apply to the Stack.
    */
-  horizontalGap?: number | string;
-
-  /**
-   * Vertical gap between Stack child elements.
-   */
-  verticalGap?: number | string;
+  styles?: IStylesProp<IStackProps, IStackStyles>;
 }
 
 export interface IStackStyles {
@@ -105,4 +99,9 @@ export interface IStackStyles {
    * Style set for the root element.
    */
   root: IStyle;
+
+  /**
+   * Style set for the inner element (only when the wrap property is true).
+   */
+  inner: IStyle;
 }