+++ packages/office-ui-fabric-react/src/components/Check/Check.styles.ts
@@ -1,5 +1,6 @@
 import { ICheckStyleProps, ICheckStyles } from './Check.types';
 import { HighContrastSelector, IStyle, getGlobalClassNames } from '../../Styling';
+import { getRTL } from '../../Utilities';
 
 const GlobalClassNames = {
   root: 'ms-Check',
@@ -11,6 +12,7 @@ export const getStyles = (props: ICheckStyleProps): ICheckStyles => {
   const { checkBoxHeight = '18px', checked, className, theme } = props;
 
   const { palette, semanticColors } = theme;
+  const isRTL = getRTL();
 
   const classNames = getGlobalClassNames(GlobalClassNames, theme);
 
@@ -107,7 +109,7 @@ export const getStyles = (props: ICheckStyleProps): ICheckStyles => {
         opacity: 0,
         color: palette.neutralTertiaryAlt,
         fontSize: '16px',
-        left: '.5px',
+        left: isRTL ? '-0.5px' : '.5px', // for centering the check icon inside the circle.
 
         selectors: {
           ':hover': {