+++ packages/office-ui-fabric-react/src/components/SelectedItemsList/SelectedPeopleList/Items/SelectedItemWithContextMenu.tsx
@@ -1,7 +1,7 @@
 /* tslint:disable */
 import * as React from 'react';
 /* tslint:enable */
-import { BaseComponent, createRef, RefObject, IBaseProps } from '../../../../Utilities';
+import { BaseComponent, IBaseProps } from '../../../../Utilities';
 import { IExtendedPersonaProps } from '../SelectedPeopleList';
 import { ContextualMenu, DirectionalHint, IContextualMenuItem } from '../../../../ContextualMenu';
 
@@ -17,7 +17,7 @@ export interface ISelectedItemWithContextMenuProps extends IBaseProps {
 }
 
 export class SelectedItemWithContextMenu extends BaseComponent<ISelectedItemWithContextMenuProps, IPeoplePickerItemState> {
-  protected itemElement: RefObject<HTMLDivElement> = createRef<HTMLDivElement>();
+  protected itemElement: React.RefObject<HTMLDivElement> = React.createRef<HTMLDivElement>();
 
   constructor(props: ISelectedItemWithContextMenuProps) {
     super(props);